requirements.txt
flask
torch
open_clip_torch
pillow
numpy
requests
